If comments option is on, still show comments despite required request header
